Move fancy version to docs dir
