+++ pyproject.toml
@@ -56,6 +56,7 @@ line-length = 100
 [tool.isort]
 profile = "black"
 line_length = 100
+known_first_party = ["uscrn", "fancy_version"]
 
 [tool.mypy]
 exclude = ["^venv/"]
